upper portion solved

+++ lib/core/model/AnimalModel.dart
@@ -224,7 +224,7 @@ class MilkDataByRow {
 
   double morningTime;
   int id;
-  int eveningTime;
+  double eveningTime;
   int sumOfOne;
   String createdOn;
